requirements.txt
beautifulsoup4>=4.12
click>=8.1
fastapi>=0.111
feedparser>=6.0
google-api-python-client>=2.0
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
markdown>=3.6
python-dotenv>=1.0
requests>=2.31
rich>=13.0
uvicorn>=0.30
playwright>=1.52
anthropic>=0.34
google-generativeai>=0.8
pytest>=8.0
